backend/requirements.txt
fastapi
uvicorn
together
pydantic
python-dotenv
pypdf
sqlalchemy
